python-fastapi/requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
SQLAlchemy==2.0.4
python-dotenv==1.0.0